Split test-amp-story-interactive-* test files to not recursively include the base test-amp-story-interactive tests in them
