+++ extensions/amp-story-interactive/0.1/test/test-amp-story-interactive-binary-poll.js
@@ -19,10 +19,7 @@ import {AmpStoryRequestService} from '../../../amp-story/1.0/amp-story-request-s
 import {AmpStoryStoreService} from '../../../amp-story/1.0/amp-story-store-service';
 import {LocalizationService} from '#service/localization';
 import {Services} from '#service';
-import {
-  addConfigToInteractive,
-  getMockInteractiveData,
-} from './test-amp-story-interactive';
+import {addConfigToInteractive, getMockInteractiveData} from './helpers';
 import {measureMutateElementStub} from '#testing/test-helper';
 import {registerServiceBuilder} from '../../../../src/service-helpers';
